This shoud be a pretty full merge, very recent updates have not yet been handled, this can be handled at school together, we must all agree it is full , everyone should run individual tests to confirm merge completion. Once this is confimed we should do a checkpoint together and determine group split and realign module configurations.
Game was not working , now fixed, issue is in how id's are being passed as objects ,
this needs to be remedied in multiple locations. JOB: araveala

A issue occurs at random intervals to do with a logger used in back end , this is the app.log file
this is only a dev tool, this will be remedied. JOB: araveala

There are multple refreshes happening in front end , they seem to alwasy call up the entire profile,
This should be investigated to conclude if neccisary or if there is a simpler alternative
(if simpler alternatives give us any benefits).

NOTE: cleanup is being left and some better comments will be added once we have our next proggresion plan.
This leaves room to make the projects morepersonal upon group split.